refactor(capabilities): migrate the wasm trampoline to the struct-hosted actor macro
Move the `#[actor(instanced)] impl NativeActor for WasmTrampoline` into
`trampoline/runtime/mod.rs` under `#[runtime]` and lift `#[actor(instanced)]`
onto the struct, per ADR-0123 (#2388). The reply-kind imports the identity's
`HandlerEntry` markers name move to the file root under the native gate.
